requirements.txt
django==1.7.7
django-debug-toolbar==1.3
django-modeltranslation==0.8
pillow==2.7
psycopg2>=2.4.5
easy-thumbnails==2.2
django-cms==3.0.12
djangocms_text_ckeditor==2.4.3
django-piwik>=0.1
flup>=1.0.2
